refactor: separate notification logic
