+++ frontend/src/components/organisms/form-sections/HeadSection/CallToProcessedMessageNotification.js
@@ -1,21 +1,52 @@
-import { useContext } from 'react';
+import { isEmpty } from 'lodash';
+import { useContext, useMemo } from 'react';
 import { markEventsAsProcessed } from '../../../../services/enrollments';
 import AlertWithTwoButtons from '../../../molecules/notification-with-buttons/AlertWithTwoButtons';
-import { FormContext } from '../../../templates/Form';
 import { OpenMessagePromptContext } from '../../../templates/Form/OpenMessagePromptContextProvider';
 import useListItemNavigation from '../../../templates/hooks/use-list-item-navigation';
+import { useAuth } from '../../AuthContext';
 
-const CallToProcessedMessageNotification = () => {
+const CallToProcessedMessageNotification = ({
+  enrollmentId,
+  events,
+  target_api,
+}) => {
   const { onClick: openMessagePrompt } = useContext(OpenMessagePromptContext);
-  const {
-    enrollment: { id },
-  } = useContext(FormContext);
   const { goBackToList } = useListItemNavigation();
+  const {
+    user,
+    user: { email },
+  } = useAuth();
 
   const markAsProcessed = async () => {
-    await markEventsAsProcessed({ id });
+    await markEventsAsProcessed({ enrollmentId });
     goBackToList();
   };
+
+  const isUserAnInstructor = useMemo(() => {
+    const targetApiInstructorRole = `${target_api}:instructor`;
+
+    return user.roles.includes(targetApiInstructorRole);
+  }, [user, target_api]);
+
+  const isEventNotifyFromDemandeur = useMemo(() => {
+    const filteredEvents = events.filter((event) => {
+      return (
+        event.name === 'notify' &&
+        event.processed_at === null &&
+        event.user.email !== email
+      );
+    });
+
+    if (isEmpty(filteredEvents)) {
+      return false;
+    }
+
+    return filteredEvents;
+  }, [events, email]);
+
+  if (!isUserAnInstructor || !enrollmentId || !isEventNotifyFromDemandeur)
+    return null;
 
   return (
     <AlertWithTwoButtons
